Rename 'state' field to 'idState' and adjust token handling
The 'state' field is renamed to 'idState' in IUser interface, UserSchema, test data and UserController for clarity and consistency. Token handling is adjusted in AuthMiddleware to check not only for empty token but also for undefined. UserController also got an interface for the update body, which explicitly requires the 'idState' key if it is provided. This is more precise and beneficial for understanding and debugging the code.